Remove bad samples from initial run.
The original from June-16-2019 did not give good results
from some unclear reason, and lacked a positive control
as we did not anticipate these problems.

Since the more recent run (July-29-2019) with a positive
control looks good, we are just discarding the June-16-2019
samples as it's hard to diagnose the problem without a positive
control.